fix: refactor LoadingApp
